Port ara_record to use the ARA 1.0 API
This port uses the (outdated) flask 1.0 as a base [1] since it had
already been switched to use the flask API and required less changes.
It otherwise doesn't change how the module itself works or the use
case for it.

The only tricky part in this implementation revolves around letting
the action plugin recover the playbook id for the playbook the plugin
is running from.
One one end, the callback saves the play uuid [2][3].
The play uuid is available to the action plugin by recursing through
its task inheritance to find the parent play's uuid.
Once the plugin has the play uuid, it is able to search for it and
recover the parent playbook's id.

[1]: https://github.com/openstack/ara/blob/0d915282e68fa1d8dc95f60840d2887d0026fc88/ara/plugins/actions/ara_record.py
[2]: https://review.openstack.org/#/c/609076/
[3]: https://review.openstack.org/#/c/609078/

Change-Id: I396ff26a53caedef7dfbd7dafa7eba9a515ea7a3
Depends-On: https://review.openstack.org/#/c/618228/