Tweak how files and hosts are created to ensure idempotency
Files and hosts can be created from different places and as such,
we need to double check if they have not already been created
whenever we might want to create them.

It should not be possible to have the same file or the same host
more than once in a given playbook due to unique constraints.

Change-Id: I6023949f4ad026086ba690789ddeebf5bb1d0a0d
Depends-On: https://review.openstack.org/#/c/625948/
Depends-On: https://review.openstack.org/#/c/625961/