Rely on the API server for the get_or_create methods
Instead of "manually" verifying if a file or a host has already
been created for the playbook, always send it and rely on the
API server to do the validation.

This is hopefully a good tradeoff: while we're now always sending
data, we save one API call whenever we want to get or create files
and hosts.

Change-Id: If16eaae9d88ed89b9e33403fe0721fab7c298143
Depends-On: https://review.openstack.org/638025